Making the "data base"
It was maked the system to store just one time the local storage.

Almost each element has the unique responsible principle.

It was maked the system to add new tasks in html and in the local storage
It was started to make system for delete the task.

I need to make the id auto adjust itself, remake it count, or find the index of an element by searching a match of the value of each property of the element